+++ src/components/Sparkbar/index.ts
@@ -1,0 +1,1 @@
+export {Sparkbar, SparkbarProps} from './Sparkbar';